# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Rishfy API
  version: 0.4.0
  description: |
    Rishfy platform API. Sprint 4: Location service, WebSocket gateway, Admin dashboard.
servers:
  - url: http://localhost/api/v1
    description: Local gateway
...
paths:
...
  /payments/initiate:
    post:
      tags: [payments]
      summary: Initiate mobile money payment for a booking
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bookingId, amountTzs, method, payerPhone, idempotencyKey]
              properties:
                bookingId: { type: string, format: uuid }
                amountTzs: { type: integer }
                method: { type: string, enum: [mpesa_tz, tigopesa, airtel_money, halopesa] }
                payerPhone: { type: string, description: 'E.164 format' }
                idempotencyKey: { type: string, format: uuid }
      responses:
        '201':
          description: Payment initiated — user receives USSD push
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentId: { type: string, format: uuid }
                  status: { type: string }
                  instructions: { type: string }
                  expiresInSeconds: { type: integer }
  /payments/{id}/status:
    get:
      tags: [payments]
      summary: Get payment status
      security: [{ bearerAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Payment status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Payment' }
  /webhooks/azampay:
    post:
      tags: [payments]
      summary: Azampay payment callback webhook
      description: 'Provider posts here on payment completion. Signature verified via HMAC-SHA256.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200': { description: Callback processed }
        '400': { description: Invalid callback payload }
        '401': { description: Invalid callback signature }
...
  /payments/{id}/refund:
    post:
      tags: [payments]
      summary: Refund a payment — cancellation policy applied
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
                departureTime: { type: string, format: date-time }
                cancelledAt: { type: string, format: date-time }
                forceFullRefund: { type: boolean, description: Admin only }
      responses:
        '200':
          description: Refund applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  paymentId: { type: string, format: uuid }
                  refundedAmountTzs: { type: integer }
                  policy: { type: string, enum: [FULL_REFUND, PENALTY_50, NO_REFUND] }
                  status: { type: string }
        '422': { description: Not refundable or already refunded }

  # ---------------------------------------------------------------------------
  # Admin dashboard
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Payment:
      type: object
      properties:
        paymentId: { type: string, format: uuid }
        status:
          type: string
          enum: [pending, completed, failed, refunded]
        providerReference: { type: string, nullable: true }
        amountTzs: { type: integer }
        refundedAmountTzs: { type: integer }
...